Install docker before trying to give it credentials.
Also, remove the explicit versions for installing apt-get packages. Also,
removed explicit versions for buildfarm images. Versioning these is not
useful, I just did that because of a linter, and while the linter is
right in most cases, these are images we build ourselves, and "latest" is
the only thing we care about here. The images should really not change
much over time anyway, so they shouldn't become incompatible.