cleanup: Merge download and setup of freebsd vm.
This avoids having 2 docker layers with a whole image in them, roughly
halving the size of the final docker image.

NB: the gzip at the end doesn't really help. It slows things down,
because docker already compresses layers, so it just adds about half a
minute to the CI build during unpacking, but changing it means changing
the toxcore build so it doesn't unpack the image, so I'm not doing it
for now.