cleanup: Speed up qtox windows builder image build.
We don't need to download the entire Qt source code. We only use 3
modules: qtbase, qttools, qtsvg. Also, parallelised the 3 main sections
of the image build: Qt, non-Qt deps, and gdb.

Also, upgraded the Windows Qt version to 6.8.1. This is possible because
of our host-qt image, which has the latest Qt built for Linux.